refactor(cli): extract root help action
