+++ src/lib/nemoclaw-runtime-bridge.ts
@@ -23,7 +23,6 @@ export interface NemoClawRuntimeBridge {
   backupAll: () => void;
   deploy: (instanceName?: string) => Promise<void>;
   garbageCollectImages: (args?: string[]) => Promise<void>;
-  help: () => void;
   onboard: (args?: string[]) => Promise<void>;
   recoverNamedGatewayRuntime: () => Promise<GatewayRecoveryResult>;
   recoverRegistryEntries: (options?: { requestedSandboxName?: string | null }) => Promise<RecoveryResult>;
@@ -49,7 +48,6 @@ export interface NemoClawRuntimeBridge {
   setup: (args?: string[]) => Promise<void>;
   setupSpark: (args?: string[]) => Promise<void>;
   upgradeSandboxes: (args?: string[]) => Promise<void>;
-  version: () => void;
 }
 
 let runtimeFactory = (): NemoClawRuntimeBridge =>
